Adds support for redirecting back to the current page in App Router (#24)
This commit adds an optional argument to the `getUserOrRedirect` function
which allows the user to specify one of:

getUserOrRedirect({ returnToPath: "/somewhere" })

or

getUserOrRedirect({ returnToCurrentPath: true })

The former is pretty straightforward. The later requires us to do some
unfortunate workarounds with Next.js, since you don't know the current URL
in server components.

Luckily, we already have the requirement that people set up middleware
due to other restrictions around whether cookies can be set in the app
router, so we can use that same pattern to pass the URL from middleware
to the downstream function.

As long as we are doing this, it also seems like a good idea to
provide a utility to other folks, so we also now expose:

`getCurrentUrl()` which has nothing to do with us, but seems generally useful.