Replace getCurrentURL for getCurrentPath
Next.js middleware can run multiple times for the same route.
This behavior, while unexpected, seems to non-deterministically
happen for our customers.

This compounds with the issue that the only way to pass information
from middleware to server component is through headers, which
users can control.

We now can't easily enforce that we are the only ones setting the
header. To alleviate this, we switch from setting the full URL,
to just setting and retrieving the path.

Thanks Next.js for the amazing product